HID: hid-sensor-hub: do not process feature reports in raw_event
In sensor_hub_raw_event(), HID feature reports are ignored but are
still marked as processed. This causes the in-kernel struct not to be
updated. Any non-updated fields in the feature reports are zero, and
they are being sent to the device. This causes confusion in the
sensor hub firmware, and some sensors are not powered up as a result.

This changes the raw_event rountine to only process input reports,
and let the hid core handle the incoming feature reports.

The issue has been observed on Acer Iconia W700.

Change-Id: I9cfb76fd48fa39d2c34bd5b956e03396bb88b729
Signed-off-by: Daniel Leung <daniel.leung@linux.intel.com>