I2C / PM_RUNTIME: enable runtime PM for i2c adapter device
The I2C device tree looks like this:

    INT33C2:0		(host device, pdev)
    /      \
   I2C-0  I2C-1		(adapter device)
    |       |
   0-0040   xxx		(client device)

Some I2C host controllers already have enabled runtime PM support, and
suppose client driver like i2c-hid will enable runtime PM for client
devices, and then we have a gap in the tree regarding runtime PM: the
adapter device I2C-x is not runtime enabled.

This would cause problem: in runtime suspend case, the host controller
would be runtime suspended even client devices are runtime active; and
in runtime resume case, when both host and client devices are runtime
suspended, runtime resume the client device will not trigger the runtime
resume of host device.

So this patch adds runtime PM support for adapter device in that when
the adapter is to be used(i.e. in i2c_transfer), runtime resume it(and
this also has the effect of runtime resume the host device); when the
transfer is finished, runtime suspend it.

Change-Id: I78511e9d69b2c5dcb493bbb33b2a19e71f505baa
Signed-off-by: Aaron Lu <aaron.lu@intel.com>
Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Signed-off-by: Kirill A. Shutemov <kirill.shutemov@linux.intel.com>