FOR_UPSTREAM [VPG]: drm/i915: Clock bending
From EDID specified dotclock is the expected clock, where as the
clock that is generated by calculation of MN P values is the actual
clock. MN P values are integers so there would be a difference
between the expected and actual. Hence the reference clock is tweaked.
This is clock bending. There is a formula to calculate the error,
which is measured in ppm. If the error is grated than 2ppm then
the bending has to be applied.
This patch implements this in order to support clock bending for BYT.
This is cpecific HDMI.

Issue: APDEV-1041
OTC-Tracker: VIZ-3447
Change-Id: Ib1ec5e42bbe0f1c1dce97b9b522f1e4bbd227605
Signed-off-by: vkannan <vandana.kannan@intel.com>
Reviewed-on: http://android.intel.com:8080/103788
Reviewed-by: cactus <cactus@intel.com>
Reviewed-by: Stimson, Dale B <dale.b.stimson@intel.com>
Tested-by: Stimson, Dale B <dale.b.stimson@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: A.Sunil Kamath <sunil.kamath@intel.com>