MODSIGN: sign-file: Add re-signing capability
Some workflows call for testing keys to be used during development
with control of production signing keys strictly controlled.
sign_file has been extended such that if it is called with a module
that already has a signature, that signature will be replaced.

This is intended to be used with the MODSIGN extensions which use
keys in BIOS DB or MokList; development devices would be provisioned
with testing keys accessible to all developers; release binaries
will be re-signed by a team that has access to the real production
keys.

For: AXIA-2117
Change-Id: I3b30af7c865bde9228883437c04d1deebd8365c6
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>
Signed-off-by: Daniel Leung <daniel.leung@intel.com>