[MODSIGN] Android: skip module checking when UEFI secure boot is OFF
Android-IA secure boot requires not to provide builtin certs, but import x509
certs in BIOS at run time. This will bring a new issue - modules signed with
BIOS keys will not pass verification if user erases the required keys in BIOS.

This patch may not be a good answer for this issue since the existence of keys
in BIOS and the status of secure boot are two different matters. Here We assume
when secure boot is enabled, the required key in BIOS shall be available too.
This is a very specific change for Android-IA.

For: AXIA-2610
Change-Id: Icd72beaff212edfcb91fad111eade9193c7eb254
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>