[MODSIGN] Disable built-in certificates in kernel
The Android-IA secure boot design requires not to bake built-in certs
when building kernel. Modules will be signed with BIOS keys and the
keys will be imported into kernel from UEFI BIOS at run time.

This patch is verified on 64-bit devices only.

Config change in 32-bit is NOT tested and just to avoid diverging.
(With other recent changes due to syncing)

For: AXIA-2610
Change-Id: I51d0ba5e7e8a77cce40002525b5561c34da25047
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>