i2c: designware: prevent signals from aborting I2C transfers
If a process receives signal while it is waiting for I2C transfer to
complete, an error is returned to the caller and the transfer is aborted.
This can cause the driver to fail subsequent transfers. Also according to
commit d295a86eab2 (i2c: mv64xxx: work around signals causing I2C
transactions to be aborted) I2C drivers aren't supposed to abort
transactions on signals.

To prevent this switch to use wait_for_completion_timeout() instead of
wait_for_completion_interruptible_timeout() in the designware I2C driver.

Signed-off-by: Mika Westerberg <mika.westerberg@linux.intel.com>
Reviewed-by: Christian Ruppert <christian.ruppert@abilis.com>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
Signed-off-by: Tushar Dave <tushar.n.dave@intel.com>

Change-Id: I181717e75ec4bbb9d3ef773084d7f0e92625f2cd
Original-Change-Id: I9735c2653f3b3843909f73a55322c525438fb677
Upstream-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Upstream-commit: e42dba569fceca5d59a88571370785e9ce9775b8
Upstream-as-of-tag: v3.12-rc1