i2c: designware: fix race between subsequent xfers
The designware block is not always properly disabled in the case of
transfer errors. Interrupts from aborted transfers might be handled
after the data structures for the following transfer are initialised but
before the hardware is set up. This can corrupt the data structures to
the point that the system is stuck in an infinite interrupt loop (where
FIFOs are never emptied because dev->msg_read_idx == dev->msgs_num).

This patch cleanly disables the designware-i2c hardware at the end of
every transfer, be it successful or not.

Signed-off-by: Christian Ruppert <christian.ruppert@abilis.com>
[wsa: extended the comment]
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>
Signed-off-by: Tushar Dave <tushar.n.dave@intel.com>

Change-Id: Ib90c12bf55e95ef894459731dfacbebdbb4e169e
Original-Change-Id: Iccf46f3a7f7ad09fbcae3f829e2f48537499db9f
Upstream-repo: git://git.kernel.org/pub/scm/linux/kernel/git/torvalds/linux.git
Upstream-commit: 38d7fadef4973bb94e36897fcb6bb6a12fdd10c9
Upstream-as-of-tag: v3.12-rc1