MUST_REBASE [VPG]: Enabling sync framework, kernel and user extensions.
Based on Gary Smith <gary.k.smith@intel.com> work for MCG.

This commit enables the user mode extensions to make use of the
sync framework within Android. This may not be needed in the longer
term as Google's preference is to provide other methods of
obtaining these file descriptors from the kernel. However, at the
moment, we dont have those alternate methods.

Adds the following to arch/x86/configs/x86_64_bigcore_android_defconfig
and i386_bigcore_android_defconfig:

CONFIG_SYNC=y
CONFIG_SW_SYNC=y
CONFIG_SW_SYNC_USER=y

Also, on the bigcore architectures, the binaries generated for most of the
components are 32 bit. This means that the ioctl calls for the sync framework
fail. The fix for this is to also register the ioctl handler in the
ioctl.

v4. Changed confusing header and commit body

Issue: AXIA-3682
Change-Id: If0e810a188f45568821f08bf5cfa02eb98c27b59
Signed-off-by: Lukasz Anaczkowski <lukasz.anaczkowski@intel.com>
Signed-off-by: Joao Santos <joao.santos@intel.com>

Conflicts:

	arch/x86/configs/i386_bigcore_android_defconfig
	arch/x86/configs/x86_64_bigcore_android_defconfig
	drivers/base/sync.c