pstore: support extra dumper
Provide new interfaces in pstore, so that other pieces in kernel
can dump their data to the persistent storage when a panic or
crash happens.

A new module parameter is also added to limit the total amount
of data from extra dumpers. It will not be reset until reboot.

For: AXIA-3125

Change-Id: I7dbf1e02200c9ab1432046901c28a2f9227e2d65
Origin-Change-Id: I1b3edd2e38e9bcccc78be71921defb89b4d12272
Signed-off-by: Jianxun Zhang <jianxun.zhang@intel.com>
Signed-off-by: Joakim Nordell <joakim.nordell@intel.com>