MUST UPSTREAM x86: boot: support minigzip bzImage compression
Android OTA system computes very efficient diffs of compressed files
if the deflate() algorithm it has access to is the same version as
used to create the original file. Here we add support for compressing
the kernel with 'minigzip' which uses the deflate() inside zlib.
This is much better than using 'gzip' as that tool has a very old
version of deflate() inside the gzip codebase instead of linking against
zlib.

Issue: AXIA-3825
Change-Id: I2b69242cfd2c96fa1e75b77113015bd6ded7f5f3
Signed-off-by: Andrew Boie <andrew.p.boie@intel.com>