REVERTME [VPG]: drm/i915: Don't switch off the eDP panel power for init sequence
This is a temporary change for BYT.
For DP/eDP aux transcations to succeed we need only the VDD.
Some of the FFRD10 PR1.1B boards (most of the board works fine without this
hack) need both the panel power & vdd to be ON for aux transactions to
succeed. This could be a board problem. Till we figure out what's going on,
use this temporary hack.

Issue:APDEV-1165
Change-Id: I5bb88f88d1f929c903185739ec3a04d50c0b6c2d
Signed-off-by: Vijay Purushothaman <vijay.a.purushothaman@intel.com>
Reviewed-on: http://android.intel.com:8080/118673
Reviewed-by: Kumar, Shobhit <shobhit.kumar@intel.com>
Tested-by: Sun, Hang L <hang.l.sun@intel.com>
Reviewed-by: buildbot <buildbot@intel.com>
Tested-by: buildbot <buildbot@intel.com>
Signed-off-by: A.Sunil Kamath <sunil.kamath@intel.com>