ACPICA: DeRefOf operator: Update to fully resolve FieldUnit and BufferField refs.
Previously, references to these objects were resolved only to the actual
FieldUnit or BufferField object. The correct behavior is to resolve these
references to an actual value.
The problem is that DerefOf did not resolve these objects to actual
values.  An "Integer" object is simple, return the value.  But a field in
an operation region will require a read operation.  For a BufferField, the
appropriate data must be extracted from the parent buffer.

NOTE: It appears that this issues is present in Windows7 but not
Windows8.

Issue: AXIA-3055
Change-Id: Ie708bf5c8218807fea3f8fb9473c3189c1bb9de8
Original-Change-Id: I4c93e5db86855d921c93e6ddfa3f0e9e7093efda
Signed-off-by: Bob Moore <robert.moore@intel.com>
Signed-off-by: Lv Zheng <lv.zheng@intel.com>
Signed-off-by: Rafael J. Wysocki <rafael.j.wysocki@intel.com>
Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>
Upstream-repo: git://git.kernel.org/pub/scm/linux/kernel/git/rafael/linux-pm.git#acpica
Upstream-commit: 63660e05ec719613b518547b40a1c501c10f0bc4
Upstream-as-of-tag: n/a (queued for 3.12)