PM: Changes in the governor to enable the Power HAL
BZ: 126718

This patch comprises governor changes that are needed to enable
the Power HAL to boost the frequency to an intermediate frequency
on user-interaction. This patch is a part of a set of 3 patches
that are used to enable Power HAL on the mainline for merrifield.
This is a rebase from http://android.intel.com:8080/#/c/123096
patch.

Change-Id: Ib226624b825fd05eb456cb033a8402203f9e556b
Signed-off-by: Ashish K <ashish.k@intel.com>
Reviewed-on: http://android.intel.com:8080/137528
Reviewed-by: Kumar P, Mahesh <mahesh.kumar.p@intel.com>
Reviewed-by: Rudramuni, Vishwesh M <vishwesh.m.rudramuni@intel.com>
Tested-by: Delance, OlivierX <olivierx.delance@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>

Conflicts:

	drivers/cpufreq/cpufreq_interactive.c

Signed-off-by: Hans Holmberg <hans.holmberg@intel.com>