xhci-plat: Don't enable legacy PCI interrupts
BZ: 129468

The xHCI platform driver calls into usb_add_hcd to register the irq for its
platform device.  It does not want the xHCI generic driver to register an
interrupt for it at all.  The original code did that by setting the
XHCI_BROKEN_MSI quirk, which tells the xHCI driver to not enable MSI or
MSI-X for a PCI host.

Unfortunately, if CONFIG_PCI is enabled, and CONFIG_USB_DW3 is enabled,
the xHCI generic driver will attempt to register a legacy PCI interrupt for
the xHCI platform device in xhci_try_enable_msi().  This will result in a bogus
irq being registered, since the underlying device is a platform_device, not a
pci_device, and thus the pci_device->irq pointer will be bogus.

Add a new quirk, XHCI_PLAT, so that the xHCI generic driver can distinguish
between a PCI device that can't handle MSI or MSI-X, and a platform device
that should not have its interrupts touched at all.
This quirk may be useful in the future, in case other corner cases like this
arise.

This patch should be backported to kernels as old as 3.9, that contain the
commit 00eed9c814cb8f281be6f0f5d8f45025dc0a97eb "USB: xhci:
correctly enable interrupts".

(cherry picked from commit 52fb61250a7a132b0cfb9f4a1060a1f3c49e5a25)

Change-Id: I3ade888ae44e63b9a6be5a192398eb947c7cc5aa
Depends-Change-Id: I4c357a98dadb0c3ae0b5ea79f001c28683d46843
Signed-off-by: Sarah Sharp <sarah.a.sharp@linux.intel.com>
Reported-by: Yu Y Wang <yu.y.wang@intel.com>
Tested-by: Yu Y Wang <yu.y.wang@intel.com>
Reviewed-by: Felipe Balbi <balbi@ti.com>
Cc: stable@vger.kernel.org
Signed-off-by: Wang, Yu <yu.y.wang@intel.com>
Reviewed-on: http://android.intel.com:8080/126666
Reviewed-by: Ranostay, Matt <matt.ranostay@intel.com>
Tested-by: Ranostay, Matt <matt.ranostay@intel.com>
Reviewed-by: cactus <cactus@intel.com>
Tested-by: cactus <cactus@intel.com>
Signed-off-by: Hans Holmberg <hans.holmberg@intel.com>

Conflicts:

	drivers/usb/host/xhci.c
	drivers/usb/host/xhci.h

Signed-off-by: Hans Holmberg <hans.holmberg@intel.com>