ACPICA: Fix for stores to field references.
Store to Arg1 (in ECWT) caused the object referred by Arg1 (CMDR) to be
overwritten by Arg0.
The change is to fix the code in exstore.c so that the store to Arg1
correctly writes to the field.  Basically, all writes to a reference that
points to a field object result in a write to the field, regardless of
whether the "ImplicitConversion" flag is set or not.
    Method (ECWT, 2, Serialized)
    {
        Store (Arg0, Arg1)
    }
    Method (ECMD, 1, Serialized)
    {
        ECWT (Arg0, RefOf (CMDR))
    }

Issue: AXIA-3055
Change-Id: Iaf81b72051810557a03de9564ba8dce19e1ec0a3
Original-Change-Id: I1b2302d11d07bb659b175f5757cab1e0731dd0c2
Signed-off-by: Bob Moore <robert.moore@intel.com>
Tested-by: Lv Zheng <lv.zheng@intel.com>
Signed-off-by: Tim Pepper <timothy.c.pepper@linux.intel.com>
Upstream-repo: pending
Upstream-commit: pending
Upstream-as-of-tag: n/a