FROM_UPSTREAM [VPG]: drm/i915: Introduce vblank work function
Issue: AXIA-4190

Throughout the driver, we have a number of pieces of code that must wait
for a vblank before we can update some state. Often these could be run
asynchronously since they are merely freeing a resource no long
referenced by a double-buffered registered. So we introduce a vblank
worker upon which we queue various tasks to be run after the next
vvblank.

This will be used in the next patches to avoid unnecessary stalls when
updating registers and for freeing resources.

References:
http://lists.freedesktop.org/archives/intel-gfx/2013-July/029795.html

v2: (Vandana)
Since the original patch had conflicts while applying on latest code base,
changes in the form of conflict resolution have been made.
Requirement:- On enabling sprite, it was found that the sprite fps was too
low. Making use of this patch as it gives a performance boost.
To be reverted:- This patch is not the final solution. The final solution
will have to be worked upon later.

Change-Id: I70984f885436a848f1cba2ae1a1847200aceb01a
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>
Signed-off-by: Vandana Kannan <vandana.kannan@intel.com>

Conflicts:

	drivers/gpu/drm/i915/intel_display.c