FROM_UPSTREAM [VPG]: drm/i915: Asynchronously unpin the old framebuffer after the next vblank
Issue: AXIA-4190

Upon completion of a modeset, we must wait for the next vblank before
releasing the old framebufer. (The scanout registers are double-buffered
and update on the next vblank. If we unpin the old scanout too soon we
run the risk of accessing invalid memory, so we first need to wait until
the scanout is reading from the new framebuffer.) As the unpinning is
then deferred, we need to make sure we hold a reference to the object
until we execute the callback.

References:
http://lists.freedesktop.org/archives/intel-gfx/2013-July/029790.html

v2: (vandana.kannan@intel.com) Rework in commit message
Requirement:- On enabling sprite, it was found that the sprite fps was too
low. Making use of this patch as it gives a performance boost.
To be reverted:- This patch is not the final solution. The final solution
will have to be worked upon later.

Change-Id: Iea934e90191112954d5d3f43c2afb9b80583371e
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>

Conflicts:

	drivers/gpu/drm/i915/intel_display.c