FROM_UPSTREAM [VPG]: drm/i915/sprite: Make plane switching asynchronous
Issue: AXIA-4190

Queue the unpinning of the current plane object to after the next vblank.
For special case benchmarks and others apps that may call set_plane at a
high frequency, we can unpin their objects directly unless they are "live".

v2: (vandana.kannan@intel.com) Rework commit message
Requirement:- On enabling sprite, it was found that the sprite fps was too
low. Making use of this patch as it gives a performance boost.
To be reverted:- This patch is not the final solution. The final solution
will have to be worked upon later.

Change-Id: I7297031fd9f3a2a2a545b1dc4df77f42cdffc5be
Signed-off-by: Jesse Barnes <jbarnes@virtuousgeek.org>
Signed-off-by: Chris Wilson <chris@chris-wilson.co.uk>

Conflicts:

	drivers/gpu/drm/i915/intel_drv.h
	drivers/gpu/drm/i915/intel_sprite.c