drm/i915: Support to create/wrap a Gem object around a User Space Ptr
A new ioctl 'DRM_IOCTL_GEM_USERPTR' has been added through which a
new GEM object can be created around a memory already allocated by
the User/App. This way a direct buffer sharing can be done with the
Graphics hardware without doing a copy to a memory allocated by
Graphics Driver. For example the buffer from a Camera App, containing
the captured Raw data of a video frame or a still image, can be directly
used by the Gfx driver for the encode.

Change-Id: I83d57133d7d526481f82dc2d34aac1fce30c5673
Signed-off-by: akashgoe <akash.goel@intel.com>