drivers/staging: Removing staging copy of SW_SYNC framework
Drivers staging has a copy of the android sync framework which was
clashing with the version in drivers/base. Since that version (drivers/
base) has it's headers file exported to userland that version is
being kept. The version is drivers/staging hasn't been removed to
provide a copy for reference.

Change-Id: Ic4d19b37b1d2f50f3c9f522b494a21f357399901
Signed-off-by: Rafael Barbalho <rafael.barbalho@intel.com>