FOR_UPSTREAM [VPG]: drm/i915: Restore the display ON/OFF state during suspend resume.
Do not turn ON the display during resume if the DPMS state is set to OFF
by the user before suspend.
In particular, for Android if the hardware composer sets the DPMS state
to OFF before suspend, then during resume do not turn ON the display.
Turn ON the display when the hardware composer sets the DPMS state to ON,
using a drm set property ioctl.

Problem description: The display turns ON on every resume, without any
user space initiated action to turn on the display like power key press.
Root cause: During resume, as soon as the software pipe prepare is done,
the driver sets the DPMS state to DPMS ON, and hence there is no check
for the DPMS state before crtc_enable. As a result the display pipe gets
enabled during resume ignoring the DPMS state during suspend.
Fix: With this patch the DPMS state is restored and a check is added to
call crtc_enable only if the DPMS state is not set to OFF during suspend.
Hence in case of android the hardware composer would set the DPMS state
to OFF just before supend and the resume will not turn on the display.
The display would turn on only after the hardware composer sets the DPMS
property to ON.

The default DPMS state remains as ON unless set to OFF by the user.
This will ensure that a suspend called from display ON state will resume to
display ON state.
If the user sets the DPMS state to OFF before supend, then the user
has to set it to ON after resume.

Issue: AXIA-4201
OTC-Tracker: AXIA-4537
Change-Id: Ie6eaf5e76b38060bf1e1f10b0143b0964f558213
Signed-off-by: vanditak <vandita.kulkarni@intel.com>