FOR_UPSTREAM [VPG]: drm/i915: Restore rps/rc6 on reset
A check of rps/rc6 state after i915_reset determined that the ring
MAX_IDLE registers were returned to their hardware defaults and that
the GEN6_PMIMR register was set to mask all interrupts. This change
restores those values to their pre-reset states by re-initializing
rps/rc6 in i915_reset. A full re-initialization was opted for versus
a targeted set of restore operations for simplicity and maintain-
ability. Note that the re-initialization is not done for Ironlake,
due to a past comment that it causes problems.

Also updated the rps initialization sequence to preserve existing
min/max values in the case of a re-init. We assume the values were
validated upon being set and do not do further range checking. The
debugfs interface for changing min/max was updated with range
checking to ensure this condition (already present in sysfs
interface).

Issue: VIZ-3142
Issue: AXIA-4676
OTC-Tracker: VIZ-3143
Change-Id: I19e7e2bd003cf1b837e85cfa4890b95a44ad1d4b
Signed-off-by: Jeff McGee <jeff.mcgee@intel.com>

Conflicts:

	drivers/gpu/drm/i915/i915_debugfs.c
	drivers/gpu/drm/i915/intel_pm.c