FOR_UPSTREAM [VPG]: drm/i915: Implement additional command checks
Check that the SDI bit is not set in MI_FLUSH_DW, specifically when
the Post-sync op is non-zero.

Allow MI_CONDITIONAL_BATCH_BUFFER_END based on a PPGTT bit check.

The other commands are currently not used by the userspace drivers and
could potentially be unsafe.

For: AXIA-4630
OTC-Tracker: AXIA-4631
Change-Id: Ied0a7fa1ec4183c04ceaa70ca36661c216d0d0cf
Signed-off-by: Brad Volkin <bradley.d.volkin@intel.com>

Conflicts:

	drivers/gpu/drm/i915/i915_reg.h