FOR_UPSTREAM [VPG]: drm/i915: Implement command parsing
At this point the parser just implements the mechanics of finding
commands in the batch buffer and looking them up in the parser tables.

It rejects poorly formed batches (e.g. no MI_BATCH_BUFFER_END command,
containing unrecognized commands, etc) but does no other checks.

Optionally enabled by a module parameter, currently defaulting to off.
We'll enable by default at the end of the series.

The code to look up commands in the per-ring tables implements a linear
search. The tables are small so in practice this does not appear to cause
a performance issue. However, the tables are sorted by command opcode such
that we can move to something like binary search if this code becomes a
bottleneck in the future.

NOTE: I found that there is a caching issue on non-LLC platforms when I
vmap() the batch buffer to access it from the kernel. It is unclear where
exactly we've got the caching out of sync. For now, we simply map the
batch buffer as write-combined (as in ion's use of vmap). This affects
performance but can be optimized in a later patch.

For: AXIA-4630
OTC-Tracker: AXIA-4631
Change-Id: If71f50d28578d27325c3438abf4b229c7cf695cd
Signed-off-by: Brad Volkin <bradley.d.volkin@intel.com>

Conflicts:

	drivers/gpu/drm/i915/Makefile
	drivers/gpu/drm/i915/i915_drv.c
	drivers/gpu/drm/i915/i915_drv.h
	drivers/gpu/drm/i915/i915_gem_execbuffer.c