FOR_UPSTREAM [VPG]: drm/i915: Integrate command parser and batch copying
With this patch, the command parser checks the batch that will
be submitted to the HW rather than the userspace buffer. There
is also an optimization in that the copy and parse operations
now share the vmap-ing of the buffer rather than each mapping
it independently. For parsing, the mapping changes from write-
combined to cached, which further improves performance.

At this point, it should be safe to revert the commit

   commit 47147b1ee85723ce7b6b45038422dc5a848fe044
   Author: mhnguyen <michael.h.nguyen@intel.com>

       [MUST_REVERT][VPG HSW-A] Make all cmdbuffers insecure

a prior workaround tracked as AXIA-3803.

[BV: Rebase onto the rest of the series, commit message, keep
setting I915_DISPATCH_SECURE when the parser is running]

For: AXIA-4496
OTC-Tracker: AXIA-4618
Change-Id: I438352efd0aa27c2b44f8d5a33aae62605d706e3
Signed-off-by: mhnguyen <michael.h.nguyen@intel.com>
Signed-off-by: Brad Volkin <bradley.d.volkin@intel.com>

Conflicts:

	drivers/gpu/drm/i915/i915_gem_execbuffer.c